requirements.txt
anthropic==0.96.0
yfinance==1.3.0
pandas==2.2.3
numpy==2.1.3
matplotlib==3.10.8
PyYAML==6.0.2
python-dotenv==1.0.1
pytest==8.3.3
